KNF, strlcpy, cleanups from kame

+++ usr.sbin/rtadvd/advcap.h
@@ -1,5 +1,5 @@
-/*	$OpenBSD: advcap.h,v 1.5 2002/02/16 21:28:08 millert Exp $	*/
-/*	$KAME: advcap.h,v 1.3 2001/02/01 09:12:08 jinmei Exp $	*/
+/*	$OpenBSD: advcap.h,v 1.6 2002/05/29 14:34:05 itojun Exp $	*/
+/*	$KAME: advcap.h,v 1.4 2001/06/08 04:46:19 jinmei Exp $	*/
 
 /*
  * Copyright (C) 1994,1995 by Andrey A. Chernov, Moscow, Russia.
@@ -38,7 +38,7 @@ __BEGIN_DECLS
 
 extern int agetent(char *, const char *);
 extern int agetflag(const char *);
-extern long long agetnum(const char *);
+extern int64_t agetnum(const char *);
 extern char *agetstr(const char *, char **);
 
 __END_DECLS
